refactor: manage tab pooling in controller
